refactor: extract otel-attributes out of telemetry
The generic OpenTelemetry attribute constructors and the error-chain
encoder lived in `telemetry`, which also carries the Sentry and PostHog
delivery stack with its own DNS bootstrap and sockets. Sans-IO crates
like `snownet` called into it for a single pure `KeyValue` constructor
and thereby linked that entire stack into every binary using them,
including the fuzz targets.

Move the pure constructors into their own crate. `snownet` no longer
depends on `telemetry` at all. `service_instance_id` stays behind,
because hashing legacy device IDs is telemetry's policy, as does
`default_resource_with`, which needs the SDK.
