+++ rust/Cargo.toml
@@ -40,6 +40,7 @@ members = [
     "libs/incremental-inet-checksum",
     "libs/known-dirs",
     "libs/logging",
+    "libs/otel-attributes",
     "libs/telemetry",
     "libs/tunnel-bypass-resolver",
     "libs/windows-security",
@@ -161,6 +162,7 @@ opentelemetry-otlp = "0.32.0"
 opentelemetry-stdout = "0.32.0"
 opentelemetry_sdk = "0.32.0"
 os_info = { version = "3.15.0", default-features = false }
+otel-attributes = { path = "libs/otel-attributes" }
 otel-instruments = { path = "libs/connlib/otel-instruments" }
 output_vt100 = "0.1.3"
 packet-coalescer = { path = "libs/connlib/packet-coalescer" }
