+++ rust/client-ffi/Cargo.toml
@@ -22,6 +22,7 @@ itertools = { workspace = true }
 libc = { workspace = true }
 logging = { workspace = true }
 opentelemetry = { workspace = true }
+otel-attributes = { workspace = true }
 otel-instruments = { workspace = true }
 phoenix-channel = { workspace = true }
 rustls = { workspace = true }
