+++ rust/client-ffi/src/platform/android/tun.rs
@@ -1,7 +1,6 @@
 use ip_packet::{IpPacket, IpPacketBuf};
 use std::os::fd::{FromRawFd, OwnedFd};
 use std::{io, os::fd::RawFd};
-use telemetry::otel;
 use tun::ioctl;
 
 pub struct Tun {
@@ -41,15 +40,15 @@ impl Tun {
         runtime.spawn(otel_instruments::periodic_queue_length(
             outbound_tx.downgrade(),
             [
-                otel::attr::queue_item_ip_packet_batch(),
-                otel::attr::network_io_direction_transmit(),
+                otel_attributes::queue_item_ip_packet_batch(),
+                otel_attributes::network_io_direction_transmit(),
             ],
         ));
         runtime.spawn(otel_instruments::periodic_queue_length(
             inbound_tx.downgrade(),
             [
-                otel::attr::queue_item_ip_packet_batch(),
-                otel::attr::network_io_direction_receive(),
+                otel_attributes::queue_item_ip_packet_batch(),
+                otel_attributes::network_io_direction_receive(),
             ],
         ));
