+++ rust/headless-client/Cargo.toml
@@ -26,6 +26,7 @@ opentelemetry = { workspace = true, features = ["metrics"] }
 opentelemetry-otlp = { workspace = true, features = ["metrics", "grpc-tonic"] }
 opentelemetry-stdout = { workspace = true, features = ["metrics"] }
 opentelemetry_sdk = { workspace = true, features = ["rt-tokio"] }
+otel-attributes = { workspace = true }
 phoenix-channel = { workspace = true }
 rpassword = { workspace = true }
 rustls = { workspace = true }
