+++ rust/libs/bin-shared/Cargo.toml
@@ -22,6 +22,7 @@ ip-packet = { workspace = true }
 ip_network = { workspace = true, features = ["serde"] }
 known-dirs = { workspace = true }
 logging = { workspace = true }
+otel-attributes = { workspace = true }
 otel-instruments = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
