+++ rust/libs/otel-attributes/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "otel-attributes"
+version = "0.1.0"
+edition = { workspace = true }
+license = { workspace = true }
+
+[dependencies]
+anyhow = { workspace = true }
+ip-packet = { workspace = true }
+opentelemetry = { workspace = true }
+uuid = { workspace = true }
+
+[dev-dependencies]
+thiserror = { workspace = true }
+tokio-tungstenite = { workspace = true }
+
+[lints]
+workspace = true
