+++ rust/libs/telemetry/Cargo.toml
@@ -10,7 +10,6 @@ bytes = { workspace = true }
 hex = { workspace = true }
 http = { workspace = true }
 http-client = { workspace = true }
-ip-packet = { workspace = true }
 moka = { workspace = true, features = ["sync"] }
 opentelemetry = { workspace = true }
 opentelemetry_sdk = { workspace = true, features = ["metrics"] }
@@ -34,7 +33,6 @@ rustls = { workspace = true, features = ["ring"] }
 socket-factory = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true, features = ["rt", "macros"] }
-tokio-tungstenite = { workspace = true }
 
 [lints]
 workspace = true
